+++ src/main/java/com/example/community_cr/community/like/repository/LikeService.java
@@ -1,0 +1,7 @@
+package com.example.community_cr.community.like.repository;
+
+public interface LikeService {
+	void likePost(long senderId, long postId);
+
+	void unlikePost(long senderId, long postId);
+}
